# OilCaseX/docs/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OilCase Api",
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/Purchased/Borehole/{id}": {
      "delete": {
        "tags": [
          "Borehole"
        ],
        "summary": "Удалять скважины можно только созданные на текущем ходу",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Accepted"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "409": {
            "description": "Уже нельзя удалить скважину",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/LithologicalData/Borehole/{x}/{y}": {
      "get": {
        "tags": [
          "Borehole"
        ],
        "summary": "Возвращает zip архив c текстовыми данными и фотографиями",
        "parameters": [
          {
            "name": "x",
            "in": "path",
            "description": "Координата сквжины  по X",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "y",
            "in": "path",
            "description": "Координата сквжины  по Y",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}